feat: atomic-design
